\name{AWAggregatorData-package}
\alias{AWAggregatorData-package}
\alias{AWAggregatorData}

\title{Data associated with AWAggregator R package}

\description{
    The `AWAggregatorData` package contains the data associated with the `AWAggregator` R package. It includes two pre-trained random forest models, one incorporating the average coefficient of variation as a feature, and the other one not including it. It also contains the PSMs in Benchmark Set 1 ~ 3 derived from the `psm.tsv` output files generated by FragPipe, which are used to train the random forest models.
}

\examples{
    \donttest{
        library(ExperimentHub)
        eh = ExperimentHub()
        myfiles = query(eh, 'AWAggregatorData')
        # myfiles[['XXX']]
    }
}
