% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPretrainedModel.R
\name{loadQuantInaccuracyModel}
\alias{loadQuantInaccuracyModel}
\title{Load a Pre-trained Random Forest Model}
\usage{
loadQuantInaccuracyModel(useAvgCV = TRUE)
}
\arguments{
\item{useAvgCV}{A logical value indicating whether the average CV attribute was used in the training.}
}
\value{
The pre-trained random forest model.
}
\description{
Loads a pre-trained random forest model for predicting the level of quantitative inaccuracy of PSMs.
}
\examples{
regr <- loadQuantInaccuracyModel(useAvgCV=TRUE)
}
