% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWAggregatorData-package.R
\docType{package}
\name{AWAggregatorData-package}
\alias{AWAggregatorData}
\alias{AWAggregatorData-package}
\title{AWAggregatorData: Attribute-Weighted Aggregation Data}
\description{
The AWAggregatorData package contains the data associated with the AWAggregator R package. It includes two pre-trained random forest models, one incorporating the average coefficient of variation as a feature, and the other one not including it. It also contains the PSMs in Benchmark Set 1~3 derived from the psm.tsv output files generated by FragPipe, which are used to train the random forest models.
}
\author{
\strong{Maintainer}: Jiahua Tan \email{jiahuatan@chem.ubc.ca} (\href{https://orcid.org/0000-0001-5839-1049}{ORCID})

Authors:
\itemize{
  \item Gian L. Negri \email{gnegri@bcgsc.ca} (\href{https://orcid.org/0000-0001-7722-8888}{ORCID})
  \item Gregg B. Morin \email{gmorin@bcgsc.ca} (\href{https://orcid.org/0000-0001-8949-4374}{ORCID})
  \item David D. Y. Chen \email{chen@chem.ubc.ca} (\href{https://orcid.org/0000-0002-3669-6041}{ORCID})
}

}
\keyword{internal}
